Vanilla 2.3 RC1

+++ forum/applications/dashboard/controllers/class.profilecontroller.php
@@ -2,7 +2,7 @@
 /**
  * Manages individual user profiles.
  *
- * @copyright 2009-2015 Vanilla Forums Inc.
+ * @copyright 2009-2016 Vanilla Forums Inc.
  * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
  * @package Dashboard
  * @since 2.0
@@ -18,6 +18,9 @@ class ProfileController extends Gdn_Controller {
 
     /** @var object User data to use in building profile. */
     public $User;
+
+    /** @var bool Can the current user edit the profile user's photo? */
+    public $CanEditPhotos;
 
     /** @var string Name of current tab. */
     public $CurrentTab;
@@ -70,7 +73,6 @@ class ProfileController extends Gdn_Controller {
         $this->ModuleSortContainer = 'Profile';
         $this->Head = new HeadModule($this);
         $this->addJsFile('jquery.js');
-        $this->addJsFile('jquery.livequery.js');
         $this->addJsFile('jquery.form.js');
         $this->addJsFile('jquery.popup.js');
         $this->addJsFile('jquery.gardenhandleajaxform.js');
@@ -88,8 +90,17 @@ class ProfileController extends Gdn_Controller {
             $this->CssClass .= 'EditMode';
         }
 
+        /**
+         * The default Cache-Control header does not include no-store, which can cause issues with outdated session
+         * information (e.g. message button missing). The same check is performed here as in Gdn_Controller before the
+         * Cache-Control header is added, but this value includes the no-store specifier.
+         */
+        if (Gdn::session()->isValid()) {
+            $this->setHeader('Cache-Control', 'private, no-cache, no-store, max-age=0, must-revalidate');
+        }
+
         $this->setData('Breadcrumbs', array());
-        $this->CanEditPhotos = c('Garden.Profile.EditPhotos') || Gdn::session()->checkPermission('Garden.Users.Edit');
+        $this->CanEditPhotos = Gdn::session()->checkRankedPermission(c('Garden.Profile.EditPhotos', true)) || Gdn::session()->checkPermission('Garden.Users.Edit');
     }
 
     /**
@@ -134,8 +145,10 @@ class ProfileController extends Gdn_Controller {
 
         $Activities = $this->ActivityModel->getWhere(
             array('ActivityUserID' => $UserID, 'NotifyUserID' => $NotifyUserIDs),
-            $Offset,
-            $Limit
+            '',
+            '',
+            $Limit,
+            $Offset
         )->resultArray();
         $this->ActivityModel->joinComments($Activities);
         $this->setData('Activities', $Activities);
@@ -169,8 +182,8 @@ class ProfileController extends Gdn_Controller {
      * @param mixed $UserID
      */
     public function clear($UserID = '') {
-        if (empty($_POST)) { // TODO: rm global
-            throw permissionException('Javascript');
+        if (!Gdn::request()->isAuthenticatedPostBack(true)) {
+            throw new Exception('Requires POST', 405);
         }
 
         $UserID = is_numeric($UserID) ? $UserID : 0;
@@ -189,29 +202,6 @@ class ProfileController extends Gdn_Controller {
             $this->jsonTarget('#Status', '', 'Remove');
             $this->render('Blank', 'Utility');
         }
-    }
-
-    /**
-     *
-     *
-     * @param $Type
-     * @param string $UserReference
-     * @param string $Username
-     * @throws Exception
-     */
-    public function connect($Type, $UserReference = '', $Username = '') {
-        $this->permission('Garden.SignIn.Allow');
-        $this->getUserInfo($UserReference, $Username, '', true);
-
-        // Fire an event and let whatever plugin handle the connection.
-        // This will fire an event in the form ProfileController_FacebookConnect_Handler(...).
-        $Connected = false;
-        $this->EventArguments['Connected'] =& $Connected;
-
-
-        $this->fireEvent(ucfirst($Type).'Connect');
-
-
     }
 
     /**
@@ -300,8 +290,8 @@ class ProfileController extends Gdn_Controller {
      * @throws Exception
      */
     public function disconnect($UserReference = '', $Username = '', $Provider) {
-        if (!$this->Request->isPostBack()) {
-            throw permissionException('Javascript');
+        if (!Gdn::request()->isAuthenticatedPostBack(true)) {
+            throw new Exception('Requires POST', 405);
         }
 
         $this->permission('Garden.SignIn.Allow');
@@ -357,7 +347,6 @@ class ProfileController extends Gdn_Controller {
         $User = Gdn::userModel()->getID($UserID, DATASET_TYPE_ARRAY);
         $this->Form->setModel(Gdn::userModel());
         $this->Form->setData($User);
-        $this->setData('User', $User);
 
         // Decide if they have ability to edit the username
         $CanEditUsername = (bool)c("Garden.Profile.EditUsernames") || Gdn::session()->checkPermission('Garden.Users.Edit');
@@ -402,6 +391,7 @@ class ProfileController extends Gdn_Controller {
             // These options become available when POSTing as a user with Garden.Settings.Manage permissions
 
             if (Gdn::session()->checkPermission('Garden.Settings.Manage')) {
+
                 // Role change
 
                 $RequestedRoles = $this->Form->getFormValue('RoleID', null);
@@ -619,7 +609,7 @@ class ProfileController extends Gdn_Controller {
         );
 
         $this->ActivityModel = new ActivityModel();
-        $Activities = $this->ActivityModel->getWhere($Where, 0, 5)->resultArray();
+        $Activities = $this->ActivityModel->getWhere($Where, '', '', 5, 0)->resultArray();
         $this->setData('Activities', $Activities);
         $this->ActivityModel->markRead(Gdn::session()->UserID);
 
@@ -702,7 +692,7 @@ class ProfileController extends Gdn_Controller {
      * @param string $Username .
      */
     public function picture($UserReference = '', $Username = '', $UserID = '') {
-        if (!c('Garden.Profile.EditPhotos', true)) {
+        if (!$this->CanEditPhotos) {
             throw forbiddenException('@Editing user photos has been disabled.');
         }
 
@@ -735,28 +725,38 @@ class ProfileController extends Gdn_Controller {
 
         if ($this->Form->authenticatedPostBack() === true) {
             $this->Form->setFormValue('UserID', $this->User->UserID);
-            $UploadImage = new Gdn_UploadImage();
-            try {
-                // Validate the upload
-                $TmpImage = $UploadImage->ValidateUpload('Picture');
 
-                // Generate the target image name.
-                $TargetImage = $UploadImage->GenerateTargetName(PATH_UPLOADS, '', true);
-                $Basename = pathinfo($TargetImage, PATHINFO_BASENAME);
-                $Subdir = stringBeginsWith(dirname($TargetImage), PATH_UPLOADS.'/', false, true);
+            // Set user's Photo attribute to a URL, provided the current user has proper permission to do so.
+            $photoUrl = $this->Form->getFormValue('Url', false);
+            if ($photoUrl && Gdn::session()->checkPermission('Garden.Settings.Manage')) {
+                if (isUrl($photoUrl) && filter_var($photoUrl, FILTER_VALIDATE_URL)) {
+                    $UserPhoto = $photoUrl;
+                } else {
+                    $this->Form->addError('Invalid photo URL');
+                }
+            } else {
+                $UploadImage = new Gdn_UploadImage();
+                try {
+                    // Validate the upload
+                    $TmpImage = $UploadImage->ValidateUpload('Picture');
 
-                // Delete any previously uploaded image.
-                $UploadImage->delete(changeBasename($this->User->Photo, 'p%s'));
+                    // Generate the target image name.
+                    $TargetImage = $UploadImage->GenerateTargetName(PATH_UPLOADS, '', true);
+                    $Basename = pathinfo($TargetImage, PATHINFO_BASENAME);
+                    $Subdir = stringBeginsWith(dirname($TargetImage), PATH_UPLOADS.'/', false, true);
 
-                // Save the uploaded image in profile size.
-                $Props = $UploadImage->SaveImageAs(
-                    $TmpImage,
-                    "userpics/$Subdir/p$Basename",
-                    c('Garden.Profile.MaxHeight', 1000),
-                    c('Garden.Profile.MaxWidth', 250),
-                    array('SaveGif' => c('Garden.Thumbnail.SaveGif'))
-                );
-                $UserPhoto = sprintf($Props['SaveFormat'], "userpics/$Subdir/$Basename");
+                    // Delete any previously uploaded image.
+                    $UploadImage->delete(changeBasename($this->User->Photo, 'p%s'));
+
+                    // Save the uploaded image in profile size.
+                    $Props = $UploadImage->SaveImageAs(
+                        $TmpImage,
+                        "userpics/$Subdir/p$Basename",
+                        c('Garden.Profile.MaxHeight', 1000),
+                        c('Garden.Profile.MaxWidth', 250),
+                        array('SaveGif' => c('Garden.Thumbnail.SaveGif'))
+                    );
+                    $UserPhoto = sprintf($Props['SaveFormat'], "userpics/$Subdir/$Basename");
 
 //            // Save the uploaded image in preview size
 //            $UploadImage->SaveImageAs(
@@ -766,22 +766,23 @@ class ProfileController extends Gdn_Controller {
 //               Gdn::config('Garden.Preview.MaxWidth', 75)
 //            );
 
-                // Save the uploaded image in thumbnail size
-                $ThumbSize = Gdn::config('Garden.Thumbnail.Size', 40);
-                $UploadImage->saveImageAs(
-                    $TmpImage,
-                    "userpics/$Subdir/n$Basename",
-                    $ThumbSize,
-                    $ThumbSize,
-                    array('Crop' => true, 'SaveGif' => c('Garden.Thumbnail.SaveGif'))
-                );
+                    // Save the uploaded image in thumbnail size
+                    $ThumbSize = Gdn::config('Garden.Thumbnail.Size', 40);
+                    $UploadImage->saveImageAs(
+                        $TmpImage,
+                        "userpics/$Subdir/n$Basename",
+                        $ThumbSize,
+                        $ThumbSize,
+                        array('Crop' => true, 'SaveGif' => c('Garden.Thumbnail.SaveGif'))
+                    );
 
-            } catch (Exception $Ex) {
-                // Throw the exception on API calls.
-                if ($this->deliveryType() === DELIVERY_TYPE_DATA) {
-                    throw $Ex;
+                } catch (Exception $Ex) {
+                    // Throw the exception on API calls.
+                    if ($this->deliveryType() === DELIVERY_TYPE_DATA) {
+                        throw $Ex;
+                    }
+                    $this->Form->addError($Ex);
                 }
-                $this->Form->addError($Ex);
             }
             // If there were no errors, associate the image with the user
             if ($this->Form->errorCount() == 0) {
@@ -815,8 +816,6 @@ class ProfileController extends Gdn_Controller {
      */
     public function preference($Key = false) {
         $this->permission('Garden.SignIn.Allow');
-
-        $this->Form->InputPrefix = '';
 
         if ($this->Form->authenticatedPostBack()) {
             $Data = $this->Form->formValues();
@@ -996,7 +995,7 @@ class ProfileController extends Gdn_Controller {
         $Result = array();
         foreach ($Data as $K => $V) {
             if (is_array($V)) {
-                $Result[$K] = self::_emoveEmailPreferences($V);
+                $Result[$K] = self::_removeEmailPreferences($V);
             } else {
                 $Result[$K] = $V;
             }
@@ -1019,19 +1018,20 @@ class ProfileController extends Gdn_Controller {
      * @access public
      * @param mixed $UserReference Unique identifier, possibly username or ID.
      * @param string $Username .
-     * @param string $TransientKey Security token.
+     * @param string $tk Security token.
      */
-    public function removePicture($UserReference = '', $Username = '', $TransientKey = '') {
+    public function removePicture($UserReference = '', $Username = '', $tk = '') {
         $this->permission('Garden.SignIn.Allow');
         $Session = Gdn::session();
         if (!$Session->isValid()) {
             $this->Form->addError('You must be authenticated in order to use this form.');
         }
 
-        // Get user data & another permission check
+        // Get user data & another permission check.
         $this->getUserInfo($UserReference, $Username, '', true);
+
         $RedirectUrl = userUrl($this->User, '', 'picture');
-        if ($Session->validateTransientKey($TransientKey) && is_object($this->User)) {
+        if ($Session->validateTransientKey($tk) && is_object($this->User)) {
             $HasRemovePermission = checkPermission('Garden.Users.Edit') || checkPermission('Moderation.Profiles.Edit');
             if ($this->User->UserID == $Session->UserID || $HasRemovePermission) {
                 // Do removal, set message, redirect
@@ -1043,8 +1043,6 @@ class ProfileController extends Gdn_Controller {
         if ($this->_DeliveryType == DELIVERY_TYPE_ALL) {
             redirect($RedirectUrl);
         } else {
-            $this->ControllerName = 'Home';
-            $this->View = 'FileNotFound';
             $this->RedirectUrl = url($RedirectUrl);
             $this->render();
         }
@@ -1109,7 +1107,7 @@ class ProfileController extends Gdn_Controller {
      * @param string $Username .
      */
     public function thumbnail($UserReference = '', $Username = '') {
-        if (!c('Garden.Profile.EditPhotos', true)) {
+        if (!$this->CanEditPhotos) {
             throw forbiddenException('@Editing user photos has been disabled.');
         }
 
@@ -1261,10 +1259,6 @@ class ProfileController extends Gdn_Controller {
                 $TabHtml = $TabName;
             }
 
-            if (!$CssClass && $TabUrl == Gdn::request()->path()) {
-                $CssClass = 'Active';
-            }
-
             $TabName = array($TabName => array('TabUrl' => $TabUrl, 'CssClass' => $CssClass, 'TabHtml' => $TabHtml));
         }
 
@@ -1329,7 +1323,7 @@ class ProfileController extends Gdn_Controller {
         $Module->addItem('Options', '', false, array('class' => 'SideMenu'));
 
         // Check that we have the necessary tools to allow image uploading
-        $AllowImages = c('Garden.Profile.EditPhotos', true) && Gdn_UploadImage::canUploadImages();
+        $AllowImages = $this->CanEditPhotos && Gdn_UploadImage::canUploadImages();
 
         // Is the photo hosted remotely?
         $RemotePhoto = isUrl($this->User->Photo);
@@ -1345,7 +1339,7 @@ class ProfileController extends Gdn_Controller {
             $Module->addLink('Options', sprite('SpDelete').' '.t('Delete Account'), '/user/delete/'.$this->User->UserID, 'Garden.Users.Delete', array('class' => 'Popup DeleteAccountLink'));
 
             if ($this->User->Photo != '' && $AllowImages) {
-                $Module->addLink('Options', sprite('SpDelete').' '.t('Remove Picture'), combinePaths(array(userUrl($this->User, '', 'removepicture'), $Session->transientKey())), array('Garden.Users.Edit', 'Moderation.Profiles.Edit'), array('class' => 'RemovePictureLink'));
+                $Module->addLink('Options', sprite('SpDelete').' '.t('Remove Picture'), userUrl($this->User, '', 'removepicture').'?tk='.$Session->transientKey(), array('Garden.Users.Edit', 'Moderation.Profiles.Edit'), array('class' => 'RemovePictureLink'));
             }
 
             $Module->addLink('Options', sprite('SpPreferences').' '.t('Edit Preferences'), userUrl($this->User, '', 'preferences'), array('Garden.Users.Edit', 'Moderation.Profiles.Edit'), array('class' => 'Popup PreferencesLink'));
@@ -1386,6 +1380,10 @@ class ProfileController extends Gdn_Controller {
             if ($this->User->Photo != '' && $AllowImages && !$RemotePhoto) {
                 $Module->addLink('Options', sprite('SpThumbnail').' '.t('Edit My Thumbnail'), '/profile/thumbnail', array('Garden.Profiles.Edit', 'Garden.ProfilePicture.Edit'), array('class' => 'ThumbnailLink'));
             }
+
+            if ($this->User->Photo != '' && $AllowImages) {
+                $Module->addLink('Options', sprite('SpDelete').' '.t('Remove Picture'), userUrl($this->User, '', 'removepicture').'?tk='.$Session->transientKey(), array('Garden.Profiles.Edit', 'Garden.ProfilePicture.Edit'), array('class' => 'RemovePictureLink'));
+            }
         }
 
         if ($this->User->UserID == $ViewingUserID || $Session->checkPermission('Garden.Users.Edit')) {
@@ -1393,7 +1391,7 @@ class ProfileController extends Gdn_Controller {
             $this->EventArguments['User'] = $this->User;
             $this->fireEvent('GetConnections');
             if (count($this->data('Connections')) > 0) {
-                $Module->addLink('Options', sprite('SpConnection').' '.t('Social'), '/profile/connections', 'Garden.SignIn.Allow');
+                $Module->addLink('Options', sprite('SpConnection').' '.t('Social'), '/profile/connections', 'Garden.SignIn.Allow', array('class' => 'link-social'));
             }
         }
     }
